Make some minot tweaks to work with release versions of KiCad
John used a nightly build of KiCad to generate the PCB for this project.
The nightly builds have some features not found in the release versions,
including some changes to the PCB file format. The main difference is the
use of the diff_pair_gap and diff_pair_width keywords, which released
versions of KiCad do not understand.

Fortunately, becuase all KiCad files are text, we can work around this by
deleting these lines. This allows the PCB file to be opened with
KiCad 4.0.6 or earlier instead of forcing peopel to use a nightly build.

The file format version number has also been adjusted from "20160815" to
"4" for similar reasons.

The original PCB file has been backed up as
badge-kicad.kicad_pcb.built_with_nightly_kicad just as a precaution.